+++ CHANGELOG.md
@@ -5,6 +5,18 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a `log_dir` configuration to the common section.
+  - If specified, logs will be written to a file named 'reproduce.log' in that
+    directory.
+  - If not specified, logs will be sent to stdout.
+  - In debug mode, logs are always sent to stdout.
+  - If `log_dir` is specified in debug mode, logs are sent to both stdout and
+    the directory.
+
 ## [0.21.1] - 2025-01-25
 
 ### Changed
@@ -357,6 +369,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   Docker, you should bind the `/report` to see the report file from the host.
 - Dockerfile changed to use g++-8
 
+[Unreleased]: https://github.com/aicers/reproduce/compare/0.21.1...main
 [0.21.1]: https://github.com/aicers/reproduce/compare/0.21.0...0.21.1
 [0.21.0]: https://github.com/aicers/reproduce/compare/0.20.1...0.21.0
 [0.20.1]: https://github.com/aicers/reproduce/compare/0.20.0...0.20.1